+++ src/Network/Network.Test/ScenarioTests/AzureFirewallTests.cs
@@ -88,6 +88,13 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-AzureFirewallCRUDwithManagementIpConfig");
         }
-        
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.azurefirewall)]
+        public void TestAzureFirewallWithDNSProxy()
+        {
+            TestRunner.RunTestScript("Test-AzureFirewallWithDNSProxy");
+        }
     }
 }